feat(extension-chrome): move out some code from utils package
