+++ packages/extension-chrome/src/pages/Notification/containers/SignData.tsx
@@ -1,6 +1,6 @@
 import React, { FC } from 'react';
 
 // TODO: implement
-export const SignBinary: FC = () => {
+export const SignData: FC = () => {
   return <div>Message signing</div>;
 };
