+++ packages/extension-chrome/src/pages/WalletManager/__tests__/mnemonic.test.ts
@@ -1,4 +1,4 @@
-import { mnemonic } from '../src';
+import { randomPickMnemonicPositions } from '../utils/mnemonic';
 const actualRandom = jest.requireActual('lodash/random');
 
 jest.mock('lodash/random', () => ({
@@ -9,14 +9,14 @@ jest.mock('lodash/random', () => ({
 it('utils#pick random position of an mnemonic', () => {
   const words = ['a', 'b', 'c', 'd', 'e', 'f', 'g', 'h', 'i', 'j', 'k', 'l'];
 
-  const positions = mnemonic.randomPickMnemonicPositions(words, 5);
-  const anotherPositions = mnemonic.randomPickMnemonicPositions(words, 5);
+  const positions = randomPickMnemonicPositions(words, 5);
+  const anotherPositions = randomPickMnemonicPositions(words, 5);
   expect(positions.size).toBe(5);
   expect(Array.from(positions).every((position) => position >= 0 && position < words.length)).toBe(true);
 
   expect(Array.from(positions)).not.toEqual(Array.from(anotherPositions));
 
-  const shortPositions = mnemonic.randomPickMnemonicPositions(words, 2);
+  const shortPositions = randomPickMnemonicPositions(words, 2);
   expect(shortPositions.size).toBe(2);
 });
 
@@ -25,6 +25,6 @@ it('utils#pack random position should invoke lodash random function', () => {
   let callCount = 2;
   mockRandom.mockImplementation(() => callCount--);
   const words = ['a', 'b', 'c'];
-  mnemonic.randomPickMnemonicPositions(words, 2);
+  randomPickMnemonicPositions(words, 2);
   expect(mockRandom).toBeCalledWith(0, 2);
 });
