+++ packages/extension-chrome/src/pages/WalletManager/utils/mnemonic.ts
