+++ packages/utils/src/index.ts
@@ -1,4 +1,3 @@
 export { LIB_VERSION } from './version';
 export * as errors from './error';
 export * as asserts from './asserts';
-export * as mnemonic from './mnemonic';
