+++ packages/utils/tsconfig.json
@@ -6,5 +6,9 @@
     "declarationMap": true,
     "outDir": "lib"
   },
-  "include": ["src"]
+  "include": [
+    "src",
+    "../extension-chrome/src/pages/WalletManager/utils/mnemonic.ts",
+    "../extension-chrome/src/pages/WalletManager/utils/mnemonic.ts"
+  ]
 }
